pyproject.toml
[tool.poetry]
name = "tg_bot_project_flask"
version = "0.1.0"
description = ""
authors = ["maryia <masha.afanaseva.29@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "8.0.1"
Flask = "2.0.1"
itsdangerous = "2.0.1"
Jinja2 = "3.0.1"
MarkupSafe = "2.0.1"
Werkzeug = "2.0.1"
psycopg2 = "^2.9.1"
Flask-SQLAlchemy = "^2.5.1"
flask-marshmallow = "^0.14.0"
python-dotenv = "^0.18.0"
marshmallow-sqlalchemy = "^0.26.1"
Flask-Migrate = "^3.0.1"
Flask-RESTful = "^0.3.9"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
